Tests/LibWeb/Text/expected/select-all-respects-user-select-none.txt

selection.toString(): "hiddenvisiblealso hiddenalso visible"
clipboard text: "visiblealso visible"
